code update

+++ WebClient/src/views/produce/Setting/ProductionLine.vue
@@ -24,7 +24,7 @@
             @click="produceLineEdit(scope.row)"
             circle
           ></el-button>
-           <el-button
+          <el-button
             icon="el-icon-delete"
             type="danger"
             size="mini"
@@ -36,13 +36,14 @@
     </el-table>
 
     <el-dialog
-      title="添加产线"
+      title="填写产线信息"
       :visible.sync="productionLineAddDialog"
       width="70%"
     >
       <el-form
         :model="editProductLine"
         status-icon
+        :rules="rules"
         ref="productionLineAddForm"
         label-width="100px"
       >
@@ -58,7 +59,7 @@
         <el-form-item label="描述" prop="description">
           <el-input v-model="editProductLine.description"></el-input>
         </el-form-item>
-        <el-form-item label="工位设置">
+        <el-form-item label="工位设置" prop="workStationList">
           <el-table
             :data="editProductLine.workStationList"
             border
@@ -152,7 +153,9 @@
           >
         </el-form-item>
         <el-form-item>
-          <el-button type="primary" @click="submit">保存</el-button>
+          <el-button type="primary" @click="submit('productionLineAddForm')"
+            >保存</el-button
+          >
           <el-button>重置</el-button>
         </el-form-item>
       </el-form>
@@ -165,6 +168,14 @@ import { mapActions } from "vuex";
 export default {
   data() {
     return {
+      rules: {
+        code: [{ required: true, message: "请输入编码", trigger: "blur" }],
+        name: [{ required: true, message: "请输入名称", trigger: "blur" }],
+        location: [{ required: true, message: "选择位置", trigger: "blur" }],
+        workStationList: [
+          { required: true, message: "请添加工位", trigger: "blur" },
+        ],
+      },
       productionLineAddDialog: false,
       employeeGroupList: [
         { groupName: "team1", groupNo: "01" },
@@ -192,7 +203,7 @@ export default {
       AddNewLineToList: "AddNewLineToList",
       FindWorkStationList: "FindWorkStationList",
       DeleteOneWorkStation: "DeleteOneWorkStation",
-      DeleteOneProduceLine:'DeleteOneProduceLine'
+      DeleteOneProduceLine: "DeleteOneProduceLine",
     }),
     addProductionLine() {
       this.productionLineAddDialog = !this.productionLineAddDialog;
@@ -206,14 +217,31 @@ export default {
     },
     addWorkStation() {
       this.editProductLine.workStationList.push({
-        indexNo: this.editProductLine.workStationList.length + 1,
-        name: "请填写名称",
+        indexNo:
+          this.editProductLine.code +
+          "-" +
+          (this.editProductLine.workStationList.length + 1),
+        name: this.editProductLine.name+(this.editProductLine.workStationList.length + 1)+'号工位',
         techId: "",
         techName: "",
         seen: false,
         employeeGroupNo: "分配组别",
         employeeGroupName: "",
       });
+    },
+    checkWorkStationList() {
+      var pass = true;
+      this.editProductLine.workStationList.map((item) => {
+        if (
+          item.produceProcessId == "" ||
+          item.produceProcessId == null ||
+          item.employeeGroupNo == "" ||
+          item.employeeGroupNo == null
+        ) {
+          pass = false;
+        }
+      });
+      return pass;
     },
     loseFcous(index, row) {
       row.seen = false;
@@ -227,22 +255,32 @@ export default {
       });
       row.produceProcessName = process.name;
     },
-    submit() {
-      this.SaveProduceLine(this.editProductLine)
-        .then((res) => {
-          if (res.resultStatus == 1) {
-            this.$message.success("保存成功");
-            this.productionLineAddDialog = !this.productionLineAddDialog;
-            if (!this.editProductLine.id) {
-              this.AddNewLineToList(res.data);
-            }
-          } else {
-            this.$message.warning(res.message);
+    submit(formName) {
+      this.$refs[formName].validate((valid) => {
+        if (valid) {
+          if (this.checkWorkStationList()) {
+            this.SaveProduceLine(this.editProductLine)
+              .then((res) => {
+                if (res.resultStatus == 1) {
+                  this.$message.success("保存成功");
+                  this.productionLineAddDialog = !this.productionLineAddDialog;
+                  if (!this.editProductLine.id) {
+                    this.AddNewLineToList(res.data);
+                  }
+                } else {
+                  this.$message.warning(res.message);
+                }
+              })
+              .catch((err) => {
+                this.$message.error(err.message);
+              });
+          }else{
+             this.$message.warning("请完善工位信息");
           }
-        })
-        .catch((err) => {
-          this.$message.error(err.message);
-        });
+        } else {
+          return false;
+        }
+      });
     },
 
     produceLineEdit(row) {
@@ -264,23 +302,41 @@ export default {
         });
     },
 
-    produceLineDelete(row){
-       this.DeleteOneProduceLine({id:row.id})
+    produceLineDelete(row) {
+      this.$messageBox
+        .confirm("确认删除产线？", "产线删除", {
+          confirmButtonText: "确定",
+          cancelButtonText: "取消",
+          type: "warning",
+        })
+        .then(() => {
+          this.DeleteOneProduceLine({ id: row.id });
+        })
+        .catch((e) => e);
     },
 
     deleteWorkStation(index, id, list) {
       if (id) {
-        this.DeleteOneWorkStation({ id: id })
-          .then((res) => {
-            if (res.resultStatus == 1) {
-              list.splice(index, 1);
-            } else {
-              this.$message.warning(res.message);
-            }
+        this.$messageBox
+          .confirm("确认删除已存在工位？", "工位删除", {
+            confirmButtonText: "确定",
+            cancelButtonText: "取消",
+            type: "warning",
           })
-          .catch((err) => {
-            this.$message.error(err.message);
-          });
+          .then(() => {
+            this.DeleteOneWorkStation({ id: id })
+              .then((res) => {
+                if (res.resultStatus == 1) {
+                  list.splice(index, 1);
+                } else {
+                  this.$message.warning(res.message);
+                }
+              })
+              .catch((err) => {
+                this.$message.error(err.message);
+              });
+          })
+          .catch((e) => e);
       } else {
         list.splice(index, 1);
       }
